% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_search.R
\name{ld_search_all}
\alias{ld_search_all}
\title{Suche über alle Kategorien hinweg (Wrapper für \code{ld_search()})}
\usage{
ld_search_all(
  term,
  pagelength = 200,
  language = "de",
  username = NULL,
  password = NULL
)
}
\arguments{
\item{term}{Suchbegriff, z. B. \code{"Bevölkerung"}.}

\item{pagelength}{Maximale Anzahl Treffer. Standard: \code{200}.}

\item{language}{Antwortsprache der API (\code{"de"} oder \code{"en"}).
Standard \code{"de"}.}

\item{username}{API-Benutzername.
Standard: Umgebungsvariable \code{LD_USERNAME}, sonst \code{"GAST"}.}

\item{password}{API-Passwort.
Standard: Umgebungsvariable \code{LD_PASSWORD}, sonst \code{"GAST"}.}
}
\description{
Diese Funktion führt eine Suche über \emph{alle} von der GENESIS API
unterstützten Kategorien aus.
}
\details{
Sie ist ein einfacher Wrapper:
\code{ld_search(term, category = "all")}.
}
\examples{
\dontrun{
ld_search_all("Migration")
}

}
