% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_search.R
\name{ld_search_tables}
\alias{ld_search_tables}
\title{Suche nach Tabellen (vereinfachte Version von \code{ld_search()})}
\usage{
ld_search_tables(
  term,
  pagelength = 200,
  language = "de",
  username = NULL,
  password = NULL
)
}
\arguments{
\item{term}{Suchbegriff, z. B. \code{"Bevölkerung"}.}

\item{pagelength}{Maximale Anzahl Treffer. Standard: \code{200}.}

\item{language}{Antwortsprache der API (\code{"de"} oder \code{"en"}).
Standard \code{"de"}.}

\item{username}{API-Benutzername.
Standard: Umgebungsvariable \code{LD_USERNAME}, sonst \code{"GAST"}.}

\item{password}{API-Passwort.
Standard: Umgebungsvariable \code{LD_PASSWORD}, sonst \code{"GAST"}.}
}
\value{
Dataframe mit zwei Spalten:
\itemize{
\item \code{code}
\item \code{content}
}
}
\description{
Diese Funktion durchsucht ausschließlich die Tabellen der
Landesdatenbank NRW und gibt nur die Spalten \code{code} und \code{content} zurück.
}
\details{
Sie ist ein bequemer Wrapper für:
\code{ld_search(term, category = "tables")}.
}
\examples{
\dontrun{
ld_search_tables("Arbeitsmarkt")
}

}
